requirements.txt
flask
werkzeug
mysql-connector-python
cryptography